Oh lawdy even more Linux+Mac fixes
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2665 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/AESEncryption.cpp
@@ -127,7 +127,7 @@ static void testWithPlaintextLength(int N)
 	std::vector<unsigned char> cyphertext = aes.encrypt(plaintext);
 
 	conPrint("encrypt took " + timer.elapsedString());
-	conPrint("cyphertext len: " + toString(cyphertext.size()));
+	conPrint("cyphertext len: " + toString((int64)cyphertext.size()));
 	conPrint("cyphertext extract: ");
 	for(int i=0; i<myMin(64, (int)cyphertext.size()); ++i)
 	{
